Extraction send-transaction assembly
