+++ src/views/chaincode/components/code.vue
@@ -132,7 +132,7 @@ import {
     addFunctionAbi,
     backgroundCompile
 } from "@/util/api";
-import transaction from "../dialog/sendTransaction";
+import transaction from "@/components/sendTransaction";
 import changeUser from "../dialog/changeUser";
 
 export default {
