+++ src/views/chaincode/oldContract.vue
@@ -70,7 +70,7 @@
 </template>
 <script>
 import contentHead from "@/components/contentHead";
-import sendTransation from "./dialog/sendTransaction"
+import sendTransation from "@/components/sendTransaction";
 import editor from "@/components/editor"
 import abiDialog from "./dialog/abiDialog"
 import { getContractList } from "@/util/api"
